requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
alembic==1.14.0
alpha_vantage==3.0.0
altair==5.5.0
annotated-types==0.6.0
anthropic==0.47.2
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.4
asgiref==3.8.1
astrapy==1.5.2
asttokens==2.4.1
async-timeout==4.0.3
asyncer==0.0.8
attrs==24.3.0
autogen==0.7.2
backcall==0.2.0
backoff==2.2.1
backtrader==1.9.78.123
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
bs4==0.0.2
build==1.2.2.post1
cachelib==0.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.3
chromadb==0.5.3
chromedriver-autoinstaller==0.6.4
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.2.0
cryptography==44.0.0
cssselect==1.2.0
cycler==0.12.1
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dirtyjson==1.0.8
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docopt==0.6.2
durationpy==0.9
et-xmlfile==1.1.0
executing==2.0.1
fast-depends==2.4.12
fastapi==0.115.2
fastjsonschema==2.19.1
feedfinder2==0.0.4
feedparser==6.0.11
ffmpy==0.3.2
filelock==3.13.1
filetype==1.2.0
finnhub-python==2.4.22
FinRobot==0.1.5
FLAML==2.3.3
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Limiter==3.9.2
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.1
flatbuffers==25.1.24
fonttools==4.49.0
fredapi==0.5.2
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.2.0
ftfy==6.3.1
geomet==0.2.1.post1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
googleapis-common-protos==1.66.0
gradio==4.19.2
gradio_client==0.10.1
greenlet==3.0.3
grpcio==1.70.0
gunicorn==20.0.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2020.1.16
html5lib==1.1
httpcore==1.0.3
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.1.2
ipykernel==6.29.2
ipython==8.31.0
itsdangerous==2.1.2
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.5
jiter==0.5.0
joblib==1.4.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
kubernetes==32.0.0
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.53
langchain-text-splitters==0.0.2
langsmith==0.1.147
limits==3.14.1
llama-index==0.10.28
llama-index-agent-openai==0.2.2
llama-index-cli==0.1.11
llama-index-core==0.10.28
llama-index-embeddings-openai==0.1.7
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.15
llama-index-multi-modal-llms-openai==0.1.5
llama-index-program-openai==0.1.5
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.16
llama-index-readers-llama-parse==0.1.4
llama-index-readers-web==0.1.8
llama-parse==0.4.0
llamaindex-py-client==0.1.16
llvmlite==0.42.0
lxml==5.3.0
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
markdownify==0.14.1
marker-pdf==0.2.6
MarkupSafe==3.0.2
marshmallow==3.21.1
matplotlib==3.10.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mmh3==5.1.0
monotonic==1.6
more-itertools==10.2.0
mplfinance==0.12.10b0
mpmath==1.3.0
msg-parser==1.2.0
msgpack==1.1.0
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
narwhals==1.21.1
nbclient==0.10.0
nbconvert==7.16.2
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.8.1
numba==0.59.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.20.1
openai==1.60.1
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
opencv-python==4.11.0.86
openpyxl==3.1.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
ordered-set==4.1.0
orjson==3.9.15
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.0.3
pandas-datareader==0.10.0
pandocfilters==1.5.1
parso==0.8.3
pdf2image==1.17.0
pdfkit==1.0.0
pdfminer.six==20240706
pdftext==0.3.20
peewee==3.17.8
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipenv==2023.12.1
pipreqs==0.5.0
platformdirs==4.3.6
playwright==1.43.0
plotly==5.19.0
posthog==3.10.0
praw==7.8.1
prawcore==2.4.0
prompt-toolkit==3.0.43
propcache==0.2.1
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyautogen==0.7.2
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
pyee==11.1.0
Pygments==2.19.1
PyJWT==2.8.0
pymongo==4.10.1
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pypandoc==1.15
pyparsing==3.1.1
pypdf==4.0.2
PyPDF2==3.0.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
pyzmq==25.1.2
RapidFuzz==3.11.0
ratelimit==2.2.1
ray==2.41.0
redis==5.0.2
referencing==0.35.1
regex==2024.11.6
reportlab==4.2.5
requests==2.31.0
requests-file==2.0.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruff==0.3.0
safetensors==0.5.2
scikit-learn==1.5.0
scipy==1.14.1
sec-api==1.0.25
selenium==4.21.0
semantic-version==2.10.0
sentence-transformers==3.0.1
sgmllib3k==1.0.0
shellingham==1.5.4
simplejson==3.19.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.37.2
streamlit==1.32.2
striprtf==0.0.26
surya-ocr==0.4.5
sympy==1.12
tabulate==0.9.0
tenacity==8.3.0
termcolor==2.5.0
texify==0.1.10
threadpoolctl==3.5.0
tiktoken==0.6.0
tinycss2==1.2.1
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.15.2
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.2
tornado==6.4.2
tqdm==4.66.1
traitlets==5.14.1
transformers==4.36.2
trio==0.25.0
trio-websocket==0.11.1
tushare==1.4.18
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.8.1
update-checker==0.18.0
urllib3==2.3.0
uuid6==2024.7.10
uvicorn==0.27.1
uvloop==0.21.0
virtualenv==20.26.2
watchfiles==1.0.4
wcwidth==0.2.13
webdriver-manager==4.0.1
webdrivermanager==0.10.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
xlrd==2.0.1
XlsxWriter==3.2.0
yarg==0.1.9
yarl==1.18.3
yfinance==0.2.52
zipp==3.21.0
